Removed file that added sample forms.  Having samples makes sense for our demo site, but I am skeptical that they are needed for clean installs.
I may add a new set later, but I’d want new samples to be in a subfolder named “examples”, and they’d need to be assigned to user_id 1.  The old samples were assigned to user_id 999999, and if that user_id didn’t exist in the database, some screens (documents tab?) would crash.